MHVN98: Add PMD interface header
Added a header to expose PMD functions for later use with the music system. Some data structures are not completely specified yet, that will come later when music is actually present. Also updated the reference START.BAT to check what soundboard is actually present, and start the appropriate PMD instance.