MHVN98: Basic graphic drawing
Character name and choices get their own boxes now (using the same graphics as the text box). And now a background can be drawn! Current background is a test pattern, but could easily in the future be replaced with a proper full image. A lot of things need to be sorted out with this system before I can call it a robust system.